backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onyx-backend"
version = "0.1.0"
description = "Project ONYX: Hybrid Sensor + CV Padel Analytics Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.20",
    "asyncpg>=0.28.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pillow>=10.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "ruff==0.2.2",
    "black==24.1.1",
    "mypy==1.8.0",
    "gunicorn==21.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "alembic*"]
